+++ server-rust/src/main.rs
@@ -30,6 +30,7 @@ const BOOT_INIT_SCRIPTS: &[&str] = &[
     "S15kvmhwd",
     "S30eth",
     "S30wifi",
+    "S40firewall",
     "S49ntp",
     "S50avahi-daemon",
     "S50sshd",